Floating point numbers should not be tested for equality
git-svn-id: http://josm.openstreetmap.de/svn/trunk@8384 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/osm/Storage.java
@@ -89,7 +89,7 @@ public class Storage<T> extends AbstractSet<T> {
     private int mask;
     private int size;
     private volatile int modCount = 0;
-    private float loadFactor = 0.6f;
+    private double loadFactor = 0.6d;
     private static final int DEFAULT_CAPACITY = 16;
     private final boolean safeIterator;
     private boolean arrayCopyNecessary;
